refactor: convert add media feature filename to keboab case
