+++ src/features/api-key/api.js
@@ -1,0 +1,5 @@
+import checkApiKey from './check-api-key';
+
+export default {
+  checkApiKey,
+};
