--- src/features/apiKey/api.js
@@ -1,5 +1,0 @@
-import checkApiKey from './checkApiKey';
-
-export default {
-  checkApiKey
-};
