+++ src/features/drop/add-media.js
@@ -1,6 +1,6 @@
 import {
   getCloudinarySignature,
-  updateDropMedia
+  updateDropMedia,
 } from '../../services/greenruhm-api/index.js';
 import { uploadToCloudinary } from '../../services/cloudinary-api/index.js';
 
@@ -8,12 +8,12 @@ export const getUploadParams = (dropId, params = {}) => {
   if (!dropId) return [];
   const fileTypes = ['posterImage', 'embedImage', 'video'];
   return fileTypes
-    .map(name => {
+    .map((name) => {
       const file = params[name] || null;
       if (!file) return;
       return { name, file, dropId };
     })
-    .filter(x => x);
+    .filter((x) => x);
 };
 
 export const mapUploadedMedia = (responses = []) =>
@@ -24,18 +24,14 @@ export const mapUploadedMedia = (responses = []) =>
         cloudinaryUrl: file.url,
         cloudinaryPublicId: file.public_id,
         cloudinaryResourceType: file.resource_type,
-        cloudinaryType: file.type
-      }
+        cloudinaryType: file.type,
+      },
     };
   }, {});
 
 const uploadImage = async ({ name, dropId, file }) => {
-  const {
-    apiKey,
-    cloudName,
-    signature,
-    timestamp
-  } = await getCloudinarySignature({ name, dropId });
+  const { apiKey, cloudName, signature, timestamp } =
+    await getCloudinarySignature({ name, dropId });
   const response = await uploadToCloudinary({
     file,
     name,
@@ -43,12 +39,12 @@ const uploadImage = async ({ name, dropId, file }) => {
     apiKey,
     cloudName,
     signature,
-    timestamp
+    timestamp,
   });
   return {
     ...response,
     name,
-    dropId
+    dropId,
   };
 };
