+++ src/features/user/api.js
@@ -1,5 +1,5 @@
 import signInUser from './signInUser';
-import requiresAuth from './requiresAuth';
+import requiresAuth from './requires-auth';
 import signUp from './signUp';
 import withMagic from './with-magic';
