+++ src/features/user/requires-auth.js
@@ -1,0 +1,8 @@
+const requiresAuth =
+  ({ errorMsg, predicate }) =>
+  (params) => {
+    if (!predicate(params)) throw new Error(errorMsg);
+    return params;
+  };
+
+export default requiresAuth;
