--- src/features/user/requiresAuth.js
@@ -1,6 +1,0 @@
-const requiresAuth = ({ errorMsg, predicate }) => params => {
-  if (!predicate(params)) throw new Error(errorMsg);
-  return params;
-};
-
-export default requiresAuth;
