+++ src/index-old.js
@@ -2,16 +2,16 @@ import drop from './features/drop/api';
 import user from './features/user/api';
 // import { checkApiKey } from './features/apiKey/api';
 import { asyncPipe, withStore } from './utils';
-import { updateApiKeyAction } from './features/apiKey/reducer';
+import { updateApiKeyAction } from './features/api-key/reducer';
 import { getUserIsSignedIn } from './features/user/reducer';
 import store from './reducer/store';
 
 const { requiresAuth } = user;
 
-const requiresSignIn = errorMsg =>
+const requiresSignIn = (errorMsg) =>
   requiresAuth({
     errorMsg,
-    predicate: params => getUserIsSignedIn(params.getState())
+    predicate: (params) => getUserIsSignedIn(params.getState()),
   });
 
 export const connect = ({ apiKey = '' }) => {
@@ -23,7 +23,7 @@ export const connect = ({ apiKey = '' }) => {
     username = '',
     title = '',
     description = '',
-    editionLimit = 0
+    editionLimit = 0,
   } = {}) =>
     asyncPipe(
       withMiddleware,
@@ -52,7 +52,7 @@ export const connect = ({ apiKey = '' }) => {
     signUp,
     signIn,
     createDrop,
-    getDrop
+    getDrop,
   };
 };
