+++ src/test/unit.js
@@ -1,5 +1,5 @@
 import '../index-test';
-import '../features/drop/addMedia-test.js';
+import '../features/drop/add-media-test.js';
 import '../features/user/reducer-test.js';
 import '../features/user/sign-up-test.js';
 import '../features/user/sign-in-test.js';
